pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"

requires = [ "setuptools>=46.1", "setuptools-scm[toml]>=5" ]

[project]
name = "pquant-ml"
description = "Pruning and Quantization of ML models"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [ { name = "NGT 1.3 team" } ]
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dynamic = [ "version" ]
dependencies = [ "hgq2", "keras>=3", "optuna", "pydantic>=2", "pyyaml>=6.0.1", "quantizers>=1.1" ]
optional-dependencies.all = [ "pytest>=8.4", "tensorflow>=2.17,<=2.20", "torch>=2.1" ]
optional-dependencies.tensorflow = [ "tensorflow>=2.17,<=2.20" ]
optional-dependencies.test = [ "pytest>=8.4" ]
optional-dependencies.torch = [ "torch>=2.1" ]
urls.repository = "https://github.com/nroope/PQuant"

[tool.setuptools]
packages = [ "pquant" ]
include-package-data = true
package-dir = { "" = "src" }

[tool.setuptools_scm]
write_to = "src/pquant/_version.py"
